requirements.txt
av==14.3.0
bcrypt==4.1.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
cryptography==42.0.7
ctranslate2==4.6.0
decorator==4.4.2
defusedxml==0.7.1
faster-whisper==1.1.1
feedparser==6.0.11
ffmpy==0.3.2
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.3.2
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.7
ifaddr==0.2.0
imageio==2.37.0
imageio-ffmpeg==0.6.0
intel-openmp==2021.4.0
Jinja2==3.1.6
librespot @ git+https://github.com/kokarare1212/librespot-python.git@2c26be215f56930721dc4f5758fce709a2b489b1
llvmlite==0.44.0
lxml==5.2.2
MarkupSafe==3.0.2
mkl==2021.4.0
more-itertools==10.7.0
moviepy==1.0.3
mpmath==1.3.0
music-tag==0.4.3
mutagen==1.47.0
networkx==3.4.2
numba==0.61.2
numpy==1.26.4
onnxruntime==1.21.1
openai-whisper==20240930
packaging==25.0
pandas==2.2.2
paramiko==3.4.0
pillow==10.3.0
proglog==0.1.12
protobuf==3.20.1
psutil==7.0.0
psycopg2==2.9.9
pwinput==1.0.3
pycparser==2.22
pycryptodomex==3.20.0
PyNaCl==1.5.0
PyOgg==0.6.14a1
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.31.0
sgmllib3k==1.0.0
six==1.16.0
sshtunnel==0.4.0
sympy==1.13.1
tabulate==0.9.0
tbb==2021.13.1
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tqdm==4.66.4
typing_extensions==4.13.2
tzdata==2024.1
urllib3==2.2.1
websocket-client==1.8.0
zeroconf==0.132.2
